+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/components/offer/create/CreateOfferViewModel.groovy
@@ -57,8 +57,6 @@ class CreateOfferViewModel {
     Boolean projectTitleValid = false
     @Bindable
     Boolean projectObjectiveValid = false
-    @Bindable
-    Boolean experimentalDesignValid = true
 
     @Bindable
     double netPrice = 0
@@ -101,6 +99,9 @@ class CreateOfferViewModel {
     @Bindable
     Boolean storageQuantityValid
 
+    @Bindable
+    EventEmitter<String> resetViewRequired
+
     Optional<Offer> savedOffer = Optional.empty()
 
     private final ResourcesService<Customer> customerResourceService
@@ -123,6 +124,7 @@ class CreateOfferViewModel {
         this.managerResourceService = managerResourceService
         this.personUpdateEvent = personUpdateEvent
 
+        resetViewRequired = new EventEmitter<>()
         offerCreatedSuccessfully = false
         this.addPropertyChangeListener("offerCreatedSuccessfully", {
             if (it.newValue as Boolean)
@@ -166,6 +168,9 @@ class CreateOfferViewModel {
         overheads = 0
         totalPrice = 0
         totalDiscountAmount = 0
+
+        projectTitleValid = false
+        projectObjectiveValid = false
 
         productItems.clear()
     }